style graph using stylesheet
https://matplotlib.org/3.2.2/tutorials/introductory/customizing.html
