+++ src/reddit_covid/__init__.py
@@ -4,6 +4,7 @@ from . import fetch
 from . import present
 from . import constants
 from . import post
+from . import graph
 
 
 def main():
@@ -11,5 +12,5 @@ def main():
     for state in constants.stateConfig.keys():
         df = fetch.fetch(state)
         stateConf = constants.stateConfig[state]
-        present.build_image(present.graph(df), stateConf)
+        present.build_image(graph.graph(df), stateConf)
         post.post(stateConf)
