+++ src/reddit_covid/present.py
@@ -2,42 +2,10 @@ from wand.image import Image
 from wand.display import display
 from wand.drawing import Drawing
 import math
-import matplotlib.pyplot as plt
 import io
 
 baseW = 504
 baseH = 384
-
-
-def graph(df):
-    posIncrease = df['positiveIncrease']
-    yticks = [
-        posIncrease.min(),
-        posIncrease.median(),
-        posIncrease.max()
-    ]
-
-    reversed = df[::-1]
-    plt.style.use('seaborn')
-    plt.stackplot(reversed['date'],
-                  reversed['positiveIncrease'], color='#d38fc5')
-    plt.tick_params(
-        axis='x',          # changes apply to the x-axis
-        which='both',      # both major and minor ticks are affected
-        bottom=False,      # ticks along the bottom edge are off
-        top=False,         # ticks along the top edge are off
-        labelbottom=False)  # labels along the bottom edge are off
-    plt.tick_params(
-        axis='y',
-        which='both',
-        labelleft=False,
-        labelright=True
-    )
-    plt.gcf().set_size_inches(4, 3)
-
-    f = io.BytesIO()
-    plt.savefig(f, bbox_inches='tight')
-    return f.getvalue()
 
 
 def build_image(graph_blob, conf):
